--- src/Package/MSTest.Sdk/Sdk/Features/DesktopTesting.targets
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-
-  <Target Name="_CheckDesktopTestingPlatform" BeforeTargets="CollectPackageReferences;Restore" Condition=" '$(TargetFramework)' != '' AND !$(TargetFramework.Contains('-windows')) ">
-    <Error Text="MSTest.DesktopTesting requires a Windows target framework (e.g. net8.0-windows). Current TargetFramework: '$(TargetFramework)'." />
-  </Target>
-
-  <ItemGroup>
-    <PackageReference Include="MSTest.DesktopTesting" Sdk="MSTest">
-      <Version Condition=" '$(ManagePackageVersionsCentrally)' != 'true' ">$(MSTestDesktopTestingVersion)</Version>
-    </PackageReference>
-    <PackageVersion Include="MSTest.DesktopTesting" Version="$(MSTestDesktopTestingVersion)"
-                    Condition=" '$(ManagePackageVersionsCentrally)' == 'true' " />
-  </ItemGroup>
-
-  <!--
-    Implicit imports
-    Ensure feature is available and user hasn't opted-out from it.
-    See https://github.com/dotnet/sdk/blob/f9fdf2c7d94bc86dc443e5a9ffecbd1962b1d85d/src/Tasks/Microsoft.NET.Build.Tasks/targets/Microsoft.NET.Sdk.CSharp.props#L26-L34
-    -->
-  <ItemGroup Condition=" '$(ImplicitUsings)' == 'true' OR '$(ImplicitUsings)' == 'enable' ">
-    <Using Include="Microsoft.MSTest.DesktopTesting" />
-  </ItemGroup>
-
-</Project>
