+++ src/Package/MSTest.Sdk/Sdk/Runner/ClassicEngine.targets
@@ -134,7 +134,7 @@
   </ItemGroup>
 
   <Import Project="$(MSBuildThisFileDirectory)../Features/Aspire.targets" Condition=" '$(EnableAspireTesting)' == 'true' " />
-  <Import Project="$(MSBuildThisFileDirectory)../Features/DesktopTesting.targets" Condition=" '$(EnableDesktopTesting)' == 'true' " />
+  <Import Project="$(MSBuildThisFileDirectory)../Features/WindowsAppTesting.targets" Condition=" '$(EnableWindowsAppTesting)' == 'true' " />
   <Import Project="$(MSBuildThisFileDirectory)../Features/Playwright.targets" Condition=" '$(EnablePlaywright)' == 'true' " />
 
 </Project>
